refactor: clean up helper components
Use same way of using classNames and rename helpText.tsx to help-text.tsx
