+++ packages/_helpers/index.ts
@@ -1,5 +1,6 @@
+export { Affix } from './affix.js';
 export { Clickable } from './clickable.js';
 export { DeadToggle } from './dead-toggle.js';
-export { Affix } from './affix.js';
 export { ExpandTransition } from './expand-transition.js';
+export { HelpText } from './help-text.js';
 export { UnstyledHeading } from './unstyled-heading.js';
