+++ packages/select/src/component.tsx
@@ -5,7 +5,7 @@ import { i18n } from '@lingui/core';
 import { label as ccLabel, select as ccSelect } from '@warp-ds/css/component-classes';
 import IconChevronDown16 from '@warp-ds/icons/react/chevron-down-16';
 
-import { HelpText } from '../../_helpers/helpText.js';
+import { HelpText } from '../../_helpers/help-text.js';
 import { activateI18n } from '../../i18n.js';
 import { useId } from '../../utils/src/index.js';
